requirements.txt
asgiref==3.8.1
attrs==24.2.0
django>=5.1.5
dj-database-url==3.0.1
django-cors-headers==4.4.0
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.5.0
drf-spectacular==0.27.2
firebase-admin==6.5.0
greenlet==3.0.3
gunicorn==23.0.0
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
Markdown==3.6
psycopg2-binary==2.9.10
PyJWT==2.9.0
PyYAML==6.0.2
referencing==0.35.1
rpds-py==0.20.0
SQLAlchemy==2.0.32
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
uvicorn==0.35.0
whitenoise==6.9.0